requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.48
alembic==1.18.4
psycopg2-binary==2.9.11
pydantic==2.12.5
python-dotenv==1.2.2
pydantic-settings==2.6.1
structlog==24.4.0
twilio==9.3.7
python-multipart==0.0.17
langgraph==1.1.3
langchain-openai==1.1.11
streamlit==1.55.0
langchain-chroma==0.1.4
langchain-community==0.3.13
pypdf==5.1.0
chromadb==0.5.23
apscheduler==3.10.4
